uint -> uint256

+++ contracts/interfaces/IMassetManager.sol
@@ -56,7 +56,7 @@ interface IMassetManager {
         address operator,
         address from,
         address to,
-        uint amount,
+        uint256 amount,
         bytes userData,
         bytes operatorData
     );
@@ -188,7 +188,7 @@ interface IMassetManager {
         address _operator,
         address _from,
         address _to,
-        uint _amount,
+        uint256 _amount,
         bytes calldata _userData,
         bytes calldata _operatorData
     ) external;
