+++ contracts/masset/MassetManager.sol
@@ -80,7 +80,7 @@ contract MassetManager is
         address operator,
         address from,
         address to,
-        uint amount,
+        uint256 amount,
         bytes userData,
         bytes operatorData
     );
@@ -478,7 +478,7 @@ contract MassetManager is
         address _operator,
         address _from,
         address _to,
-        uint _amount,
+        uint256 _amount,
         bytes calldata _userData,
         bytes calldata _operatorData
     ) external {
