+++ contracts/mocks/shared/MockERC20.sol
@@ -27,7 +27,7 @@ contract MockERC20 is ERC20 {
         address operator,
         address from,
         address to,
-        uint amount,
+        uint256 amount,
         bytes calldata userData,
         bytes calldata operatorData
         ) external {
